+++ src/app/(app)/members/components/MemberFormBuilder.tsx
@@ -13,7 +13,7 @@ import { Input } from "@/components/ui/input";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { z } from "zod";
-import { MemberRole, Shift } from "@/constant/constant";
+import { MemberRole, SHIFT_TIMING, Shift } from "@/constant/constant";
 import {
   Select,
   SelectContent,
@@ -55,24 +55,6 @@ export const USER_ROLES = [
   },
 ];
 
-export const SHIFT_TIMING = [
-  {
-    time: "Shift 12:00am to 00:06am",
-    shift: Shift.ShiftA,
-  },
-  {
-    time: "Shift 00:06am to 00:12pm",
-    shift: Shift.ShiftB,
-  },
-  {
-    time: "Shift 00:12pm to 00:06pm",
-    shift: Shift.ShiftC,
-  },
-  {
-    time: "Shift 00:06pm to 00:12am",
-    shift: Shift.ShiftD,
-  },
-];
 
 export const userBuilderSchema = z.object({
   name: z.string().min(1, {
