+++ src/app/api/cron/attendance/route.ts
@@ -1,98 +1,69 @@
 import { Tables } from "@/lib/db";
 import { getSupabaseClient } from "@/utils/supabase/supabaseClient";
 import { NextResponse } from "next/server";
-import { AttendanceStatus, MemberRole, Shift } from "@/constant/constant";
+import {
+  AttendanceStatus,
+  MemberRole,
+  SHIFT_TIMING,
+} from "@/constant/constant";
 import * as cron from "node-cron";
-
-const SHIFT_TIMING = [
-  {
-    time: "Shift 12:00am to 00:06am",
-    shift: Shift.ShiftA,
-  },
-  {
-    time: "Shift 00:06am to 00:12pm",
-    shift: Shift.ShiftB,
-  },
-  {
-    time: "Shift 00:12pm to 00:06pm",
-    shift: Shift.ShiftC,
-  },
-  {
-    time: "Shift 00:06pm to 00:12am",
-    shift: Shift.ShiftD,
-  },
-];
 
 export async function GET() {
   try {
     cron.schedule("0 0 0 * * *", async () => {
-      // Get All Attendances
       const supabase = await getSupabaseClient();
-      const { data: attendance, error } = await supabase
+
+      const { data: allAttendance, error } = await supabase
         .from(Tables.Attendance)
         .select(`*,${Tables.Members} (role,shift,phoneNumber)`)
         .eq("status", AttendanceStatus.Pending);
 
+      // Get All Attendances and Incharge Phone Number
       if (error) {
         throw error;
       }
       // In case of Holidays
-      else if (attendance.length === 0) {
-        return attendance;
+      if (allAttendance?.length === 0) {
+        return;
       }
-      // Get Incharge Phone Number from the database
-      const { data: incharge, error: inchargeError } = await supabase
-        .from(Tables.Members)
-        .select("phoneNumber")
-        .eq("role", MemberRole.Incharge);
-
-      if (inchargeError) {
-        throw inchargeError;
-      }
-      //Send Notification to Incharge
-      console.warn(incharge);
 
       // Filter All Shift Incharges Attendance
-      const shiftInchargesAttendance = attendance?.filter(
-        (attendance) => attendance?.members.role === MemberRole.ShiftIncharge
+      const shiftInchargesAttendance = allAttendance?.filter(
+        (attendance) => attendance.members.role === MemberRole.ShiftIncharge
       );
 
-      // Filter All Members
-      const memberAttendance = attendance?.filter(
-        (attendance) => attendance.members.role === MemberRole.Member
+      //Send Notification to Incharge
+      const inchargeAttendance = allAttendance?.find(
+        (attendance) => attendance.members.role === MemberRole.Incharge
       );
+      // handle Incharge Notification
+
+      // in case of absence of incharge then we will go to member table for getting phone number of the incharge
+      if (!inchargeAttendance) {
+        await supabase.from(Tables.Members).select("phoneNumber");
+        // handle Incharge Notification
+      }
+
       //Send Notification to Shift Incharge
-      SHIFT_TIMING.map(async (shiftIncharge) => {
-        const shiftInchargePhoneNumber = shiftInchargesAttendance.find(
-          (attendance) => attendance.members.shift === shiftIncharge.shift
+      for (let shiftTiming of SHIFT_TIMING) {
+        const shiftInchargePhoneNumber = shiftInchargesAttendance?.find(
+          (attendance) => attendance.members.shift === shiftTiming.shift
         );
-
         if (shiftInchargePhoneNumber) {
-          // send notification to the phone number of the shift Incharge
-          return shiftInchargePhoneNumber.members.phoneNumber;
+          // Handle Whatsapp Notification for  attended Shift Incharge
+          return;
         }
-        const { data: phoneNumberOfunattendedShiftIncharge, error } =
-          await supabase
-            .from(Tables.Members)
-            .select("phoneNumber")
-            .eq("role", MemberRole.ShiftIncharge)
-            .eq("shift", shiftIncharge.shift);
+        const { error } = await supabase
+          .from(Tables.Members)
+          .select("phoneNumber")
+          .eq("role", MemberRole.ShiftIncharge)
+          .eq("shift", shiftTiming.shift);
         if (error) {
           throw error;
         }
-        // send Notification to the phone number of unattended shiftIncharge
-        return phoneNumberOfunattendedShiftIncharge;
-      });
-      // Map members to related shift
-      SHIFT_TIMING?.map((shift) => {
-        const results = memberAttendance.filter(
-          (attendance) => attendance.members.shift === shift.shift
-        );
-        return {
-          shift: shift.shift,
-          attendance: results,
-        };
-      });
+        // Handle Whatsapp Notification for Un attended Shift Incharge
+        return;
+      }
     });
     return NextResponse.json({ data: "Success", status: 200 });
   } catch (error) {
